requirements.txt
Update:

Pillow==9.3.0

